♻️ refactor: reorganize folder structure for better modularity
- Moved component-specific logic from utils to components folder
- Kept only shared utility functions in the utils folder
